% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BETOS_TB}
\alias{BETOS_TB}
\title{BETOS Codes}
\format{An object of class \code{data.frame} with 106 rows and 2 columns.}
\source{
documentation for the record layout of HCPCS 2019 record layout file.
}
\usage{
BETOS_TB
}
\description{
Codes and Descriptions for the Berenson-Eggers Type of Services (BETOS) for
the procedure code based on generally agreed upon clinically meaningful
groupings of procedures and services.
}
\keyword{datasets}
