% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CMS_TYPE_SRVC_TB}
\alias{CMS_TYPE_SRVC_TB}
\title{CMS Type Service Codes}
\format{An object of class \code{data.frame} with 34 rows and 2 columns.}
\source{
documentation for the record layout of HCPCS 2019 record layout file.
}
\usage{
CMS_TYPE_SRVC_TB
}
\description{
CMS Type Service Codes
}
\keyword{datasets}
