% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HCPCS_2019_Q4}
\alias{HCPCS_2019_Q4}
\title{HCPCS (Including CPT)}
\format{An object of class \code{data.frame} with 12253 rows and 5 columns.}
\source{
correctcodechek.decisionhealth.com
}
\usage{
HCPCS_2019_Q4
}
\description{
HCPCS (Including CPT)
}
\keyword{datasets}
