% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HCPCS_2019}
\alias{HCPCS_2019}
\title{HCPCS Contractor Record 2019}
\format{An object of class \code{data.frame} with 11622 rows and 26 columns.}
\usage{
HCPCS_2019
}
\description{
2019 Healthcare Common Procedure Coding System (HCPCS) Contractor record
description.
}
\details{
The information provided here comes from the HCPCS Record Layout files
provided by the Centers for Medicare and Medicaid Services.

\describe{
  \item{HCPCS_CD}{Healthcare Common Procedure Coding System Code}
}
}
\seealso{
vignette("hcpcs", package = "cms.hcpcs")
}
\keyword{datasets}
